microservices/video-generation-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-storage==2.13.0
google-cloud-logging==3.8.0
google-cloud-texttospeech==2.14.1
google-generativeai==0.3.1
moviepy==1.0.3
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.2
ffmpeg-python==0.2.0
websockets==12.0
uuid==1.30 